cleanup: move workload patching to a subpackage
